+++ .gitignore
@@ -37,5 +37,3 @@ out/
 .vscode/
 
 
-### application.yml
-*.yml
