+++ src/test/java/shopping/infrastructure/config/CurrencyLayerExchangeRateApiTestConfig.java
@@ -1,33 +1,47 @@
 package shopping.infrastructure.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import java.time.Duration;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Profile;
 import org.springframework.web.client.RestTemplate;
-import shopping.config.TestConfig;
 import shopping.infrastructure.CurrencyLayerExchangeRateApi;
 import shopping.infrastructure.ExchangeRateApi;
 
 @Configuration
-@Import(TestConfig.class)
+@Profile("currencyTest")
 public class CurrencyLayerExchangeRateApiTestConfig {
 
-    @Autowired
-    private RestTemplate restTemplate;
-
+    private static final int CONNECTION_TIME_OUT = 5000;
+    private static final int READ_TIME_OUT = 5000;
     @Value("${currency.accessKey}")
     private String accessKey;
     @Value("${currency.baseUrl}")
     private String baseurl;
 
     @Bean
+    public RestTemplateBuilder restTemplateBuilder() {
+        return new RestTemplateBuilder();
+    }
+
+
+    @Bean
+    public RestTemplate restTemplate(final RestTemplateBuilder restTemplateBuilder) {
+
+        return restTemplateBuilder
+            .setConnectTimeout(Duration.ofMillis(CONNECTION_TIME_OUT))
+            .setReadTimeout(Duration.ofMillis(READ_TIME_OUT))
+            .build();
+    }
+
+    @Bean
     public ExchangeRateApi exchangeRateApi() {
         return new CurrencyLayerExchangeRateApi(
             accessKey,
             baseurl,
-            restTemplate
+            restTemplate(restTemplateBuilder())
         );
     }
 }
