+++ src/test/resources/application.yml
@@ -1,0 +1,26 @@
+spring:
+  jpa:
+    database-platform: org.hibernate.dialect.H2Dialect
+    hibernate:
+      ddl-auto: none
+    properties:
+      hibernate:
+        format_sql: true
+        show_sql: true
+    defer-datasource-initialization: true
+  sql:
+    init:
+      mode: always
+      data-locations: classpath:test-data.sql
+  profiles:
+    active: test
+
+security:
+  jwt:
+    token:
+      secret-key: 66eI65287YOV64SI66y066eb7J6I7Ja07JqU66eo64Kg66i57J2E7IiY7J6I7Ja0
+      expire-length: 3600000
+      
+currency:
+  accessKey: 1234
+  baseUrl: "http://api.currencylayer.com"
